requirements.txt
Django==5.1.4
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
requests==2.31.0